changed routes and started working on the welcome page
