+++ client/src/components/NavComponent.vue
@@ -2,14 +2,17 @@
 <template>
   <nav>
     <div class="logoname">
-      <img src="../assets/travel-ver2.svg" style="height: 70px" />
-      <h2>Travel.Talk</h2>
+      <img src="../assets/svgviewer-output.svg" style="height: 70px" />
+      <h2>TravelTalk</h2>
     </div>
 
     <div class="ul-container">
       <ul>
-        <li><a href="">Profile</a></li>
-        <li><a href="">About</a></li>
+        <li>
+          <router-link to="/">Home</router-link>
+        </li>
+        <li><router-link to="/post">Post</router-link></li>
+        <li><router-link to="/signup">Account</router-link></li>
 
         <button
           class="navBtn"
@@ -27,9 +30,9 @@
 export default {
   name: "navComponent",
   async created() {
-    if (localStorage.getItem("token") === null) {
-      this.$router.push("/");
-    }
+    // if (localStorage.getItem("token") === null) {
+    //   this.$router.push("/");
+    // }
   },
 
   methods: {
@@ -55,12 +58,13 @@ nav {
   position: fixed;
   top: 0;
   width: 100%;
-  font-size: 20px;
+  height: 9%;
 }
 
 .ul-container {
   display: flex;
   justify-content: space-between;
+  width: 300px;
   margin-right: 2rem;
 }
 li {
@@ -70,15 +74,18 @@ ul > li > a {
   text-decoration: none;
   color: #f5f7f8;
 }
-ul > li > a:hover {
-  color: #5f7e74;
+ul > li:hover,
+ul > .router-link-active,
+.router-link-exact-active {
+  border-bottom: 5px solid var(--landingClr);
   font-weight: bolder;
 }
 ul {
   display: flex;
   justify-content: space-between;
   align-items: center;
-  width: 200px;
+  width: 100%;
+  font-size: 1.3rem;
 }
 .logoname {
   display: flex;
@@ -90,8 +97,19 @@ ul {
 .navBtn {
   border: none;
   border-radius: 10px;
-  height: 2rem;
-  background-color: #45474b;
+  height: 3rem;
+  font-size: 1.2rem;
+  width: 40%;
+  background-color: var(--landingClr);
   color: #f5f7f8;
+  font-weight: bold;
+  transition: ease-in-out 0.3s;
+}
+
+.navBtn:hover {
+  background-color: var(--backgroundClr);
+  color: var(--landingClr);
+  border: 3px solid var(--landingClr);
+  transition: ease-in-out 0.3s;
 }
 </style>
